refactor: Update image layout in showBlog, showKomunitas, and showProduk views
This commit refactors the image layout in the showBlog.blade.php, showKomunitas.blade.php, and showProduk.blade.php views. It replaces the previous layout with a new layout that includes a card-img-wrapper div with a width of 75% and a mx-auto class for center alignment. Additionally, it adds a d-flex justify-content-center div to center the image within the card-img-wrapper. This change improves the visual presentation of the images in the respective views and ensures consistent styling across the application.